Adicionado seeder para pontos

+++ database/migrations/2019_01_30_000005_create_scores_table.php
@@ -16,7 +16,7 @@ class CreateScoresTable extends Migration
         Schema::create('scores', function (Blueprint $table) {
             $table->increments('id');
             $table->string('title')->comment('Título do ponto');
-            $table->float('value')->comment('Valor do ponto');
+            $table->double('value')->comment('Valor do ponto');
             $table->timestamps();
         });
     }
